Allow escapes of {, }, (, and ), and wholly use such unescaped for parsing groups and variable declarations.
Long group regexes are no longer truncated in the 'Test' command output - this previously hid the issue that is fixed in this commit.

Fix to work with non-trivial group layouts. Note that the output given by the 'Test' command is slightly wrong for strings containing non-ascii characters, as the length of the string in bytes is larger than the length of the string itself, but the underlying value in the pre-processor is accurate as it counts the unicode properly - this can be demonstrated by running the optimiser.